[#3232 #3234] Update management command
Add --live option. Running without it only lists changes that would be made, but does not change anything.

Add assertions to abort when conditions are too bad to continue.

+++ akvo/rsr/management/commands/single_period_indicators.py
@@ -9,7 +9,10 @@ import tablib
 from django.conf import settings
 from django.core.management.base import BaseCommand
 
+from optparse import make_option
+
 from akvo.rsr.models import IndicatorPeriodData
+from akvo.utils import single_period_dates
 from ...models import Project, Indicator, IndicatorPeriod
 
 
@@ -17,22 +20,43 @@ class Command(BaseCommand):
     args = '<name> <all | ids of projects>'
     help = ('Convert a project hierarchy to use single period reporting')
 
+    option_list = BaseCommand.option_list + (
+        make_option('--live',
+                    action='store_true',
+                    dest='live',
+                    default=False,
+                    help='Live run, changes made to the models'),
+    )
+
     def handle(self, *args, **options):
 
         def set_period_dates(period):
-            period.period_start = start_date
-            period.period_end = end_date
+            period.period_start = START_DATE
+            period.period_end = END_DATE
             period.save()
+
+        def get_parent_periods(indicator, START_DATE, END_DATE):
+            return indicator.parent_indicator.periods.filter(
+                    period_start=START_DATE, period_end=END_DATE)
 
         name = args[0]
         config = settings.SINGLE_PERIOD_INDICATORS[name]
-        start_date, end_date = config['start_date'], config['end_date']
+        live = options['live']
+
+        START_DATE, END_DATE = single_period_dates(name)
+        assert START_DATE, u"No start date configured. Aborting."
+        assert END_DATE, u"No end date configured. Aborting."
+
+        pk = config['pk']
+        root = Project.objects.get(pk=pk)
+        hierarchy_projects = root.descendants()
+
         if args[1] == 'all':
-            pk = config['pk']
-            root = Project.objects.get(pk=pk)
-            projects = root.descendants()
+            projects = hierarchy_projects
         else:
             projects = Project.objects.filter(id__in=map(int, args[1:]))
+            in_hierarchy = set(projects) < set(hierarchy_projects)
+            assert in_hierarchy, u"Not all projects part of the hierarchy. Aborting"
 
         data = tablib.Dataset(
             headers=[
@@ -48,7 +72,7 @@ class Command(BaseCommand):
             for indicator in Indicator.objects.filter(
                     result__project=project).select_related('result'):
                 periods = IndicatorPeriod.objects.filter(
-                    indicator=indicator).order_by('-period_end')
+                        indicator=indicator).order_by('-period_end')
                 period_count = periods.count()
                 row_data = [
                     project.pk,
@@ -57,29 +81,59 @@ class Command(BaseCommand):
                     indicator.pk,
                     indicator.title.replace('\t', ' ').replace('\n', ' '),
                 ]
+
+                # No period present. Create one and link it if indicator has a parent
                 if period_count == 0:
-                    period = IndicatorPeriod.objects.create(
-                        indicator=indicator,
-                        period_start=start_date,
-                        period_end=end_date
-                    )
-                    row_data += ['Created period: {}'.format(period.pk)]
+                    # If the indicator has a parent we need to find and link to the parent period
+                    if indicator.parent_indicator:
+                        parent_periods = get_parent_periods(indicator, START_DATE, END_DATE)
+                        if parent_periods.count() != 1:
+                            row_data += ["ERROR: Can't find parent period."]
+                            continue
+                        else:
+                            parent_period = parent_periods[0]
+                    else:
+                        parent_period = None
+                    if live:
+                        period = IndicatorPeriod.objects.create(
+                            indicator=indicator,
+                            parent_period=parent_period,
+                            period_start=START_DATE,
+                            period_end=END_DATE,
+                        )
+                        row_data += ['Created period: {}'.format(period.pk)]
+                    else:
+                        row_data += ['Would create period']
+
+                # One period. Change its dates.
                 elif period_count == 1:
                     period = periods[0]
-                    set_period_dates(period)
-                    row_data += ['Modified period: {}'.format(period.pk)]
+                    if live:
+                        set_period_dates(period)
+                        row_data += ['Modified period: {}'.format(period.pk)]
+                    else:
+                        row_data += ['Would modify period: {}'.format(period.pk)]
+
+                # Multiple periods. Keep the "latest" one.
                 else:
                     updates = IndicatorPeriodData.objects.filter(period__in=periods)
                     if updates.count() > 0:
-                        row_data += ['Multiple periods, at least one with updates']
+                        row_data += ['ERROR: Multiple periods, at least one with updates']
                     else:
-                        for period in periods[1:]:
-                            period.delete()
-                        set_period_dates(periods[0])
-                        row_data += ['Deleted {} periods, set dates on period {}'.format(
-                            period_count - 1, periods[0]
-                        )]
+                        if live:
+                            for period in periods[1:]:
+                                period.delete()
+                            set_period_dates(periods[0])
+                            row_data += ['Deleted {} periods, set dates on period {}'.format(
+                                period_count - 1, periods[0].pk
+                            )]
+                        else:
+                            row_data += ['Would delete {} periods, set dates on period {}'.format(
+                                period_count - 1, periods[0].pk
+                            )]
 
                 data.append(row_data)
+
         # Exporting as tsv results in errors I can't explain in som cases, columns get merged :-(
+        # Looks like a bug in tablib
         print data.export('csv')
